Move into `InputText`
