--- packages/components/src/InputTemplate/InputTemplate.css
@@ -1,9 +1,0 @@
-.inputTemplate {
-  --message-weight: 400;
-
-  font-weight: var(--message-weight);
-}
-
-.bold {
-  --message-weight: 800;
-}
