--- packages/components/src/InputTemplate/InputTemplate.css.d.ts
@@ -1,2 +1,0 @@
-export const inputTemplate: string;
-export const bold: string;
