--- packages/components/src/InputTemplate/InputTemplate.mdx
@@ -1,65 +1,0 @@
----
-name: InputTemplate
-menu: Components
-route: /components/input-template
----
-
-import { Playground, Props } from "docz";
-import { ComponentStatus } from "@jobber/docx";
-import { useState, useEffect } from "react";
-import { Button } from "../Button";
-import { InputTemplate } from ".";
-
-# InputTemplate
-
-<ComponentStatus stage="pre" responsive="no" accessible="no" />
-
-In Jobber a InputTemplate is used to …
-
-## Simple Example
-
-<Playground>
-  {() => {
-    const [title, setTitle] = useState("Bob Bobbed us.");
-    const [text, setText] = useState("Bob Bobbed us.");
-    {/* const [inserter, setInserter] = useState(); */}
-    const reset = () => {
-      setText("Bob Bobbed us.");
-    };
-
-    let inserter = undefined;
-    const registerInserter = newInserter => {
-      inserter = newInserter
-      {/* setInserter(newInserter); */}
-    }
-
-    const insert = () => {
-      inserter && inserter.current && inserter.current("Bar");
-    }
-    const handleTitleChange = newValue => {
-      setTitle(newValue);
-    };
-    const handleTextChange = newValue => {
-      setText(newValue);
-    };
-
-    return (
-      <>
-        <InputTemplate value={title} onChange={handleTitleChange} registerInserter={registerInserter}  />
-        <InputTemplate value={text} onChange={handleTextChange} registerInserter={registerInserter}  />
-        <pre>
-          {title}<br />
-          {text}
-        </pre>
-        <Button label="Reset" onClick={reset} />
-        <Button label="Insert 2" onClick={insert} />
-      </>
-    );
-
-}}
-
-</Playground>
-
-## InputTemplate Properties
-
-<Props of={InputTemplate} />
